extract: normalize modes like create
This causes extract to ignore the exact mode in a tarball and only
consider the same parts significant that create (and git) do. This
means that the mode is completely ignored for directories and
symlinks; for plain files, the mode is 0o755 if the user executable
bit is set and 0o644 otherwise.