allow write into directory overwriting a symlink
I realized that a path should be removed from the links set if it
a non-link overwrites a link. This fixes that logic and adds a test
that exercises this issue:

- create `path` as a symlink
- overwrite `path` as a directory
- write `path/file` inside of that directory

Previously this triggered the "symlink ovewrite attack" detection
since `path` was added to the links set and not removed when the
symlink was replaced by a plain directory. Now `path` is removed
from the links set when the symlink is replaced by a directory, so
the write to `path/file` is allowed.

Admittedly, this is a weird thing for a tarball to do, but since it
is safe, we may as well allow it for the sake of consistency.