write_tarball: reduce the number of lstat calls
We were lstat'ing each file twice before: once to see if it's a
directory when sorting entries to emit the tarball and again to
generate a tar header for it. This saves the first lstat result
and passes it into the recursive call to write_tarball, thereby
avoiding the second lstat call.