write_data: rewrite; only require full read of files
For IO streams, we might want to read some data in the middle of
a larger chunk of data. To be fully useful, this also requires being
able to sync the stream to the right location at read start, which
is provided by passing (IO, Integer) as the data argument.