read_data without sink: return view into buffer
Previously read_data without an IO handle or file to write data to
read data from the tarball into a buffer and then into a string via
an IOBuffer. This change expands the buffer if necessary and then
returns a view into the buffer. Thereby avoiding an unnecessary
copy of the data. The buffer is expanded to the next biggest power
of two since otherwise it's likely that reading data by multiples
of the buffer size will be less efficient.