open_write(IO): always flush IO handle
Any time we open an IO handle for writing, make sure to flush it
before returning from the API function.

Test by testing that full output is generated for calls that write to
IO handles even before the IO handle is closed. This includes
`Tar.create` and `Tar.rewrite`. In the process of testing this, I
noticed that the `Tar.create` tests that were supposed to be testing
writing to an IO handle were accidentally writing to a file instead,
so I fixed that first, otherwise some of these test changes would not
have caused failures without the primary change in this commit.