use ArgTools for read/write args & testing
The new ArgTools package abstracts out the open_{read,write} functionality we
had in here and adds tooling for comprehensive testing. In this first commit, we
leave the functionality of the Tar package unchanged--the same set of argument
types are accepted. In a separate commit, we can unrestrict this and get support
for command arguments and pipelines, for free (and tests).