Fix padding assertion errors
Tar formats pad their inputs with zeros to have all content aligned to
a certain block size; when extracting data, we used a `while` loop to
iterate until we had consumed all non-padded data within the stream,
when we actually needed to iterate until we had consumed all of the
zero-padded data as well.  This was only an issue when the stream
buffering lined up such that a `readbytes!()` call ends with reading all
nonzero content, but not all zero-padding.

To test this, I have added a buffer stream type that allows us to
control the number of bytes returned from a read operation, as well as
allowing it to be randomized, for some simple fuzzing.