Tar.extract: slight mode improvement
The reason for using `filemode` to read the created mode was to respect
the umask or equivalent mechanism for default permissions. Since the
umask can only take permissions away, it is always correct to AND the
default permissions with the desired permissions, since that will be
more constrained than both the default and what the tarball requests.