fix read_data bug & use in read_standard_header
Also make sure that buffer views are constructed with `Int` ranges.
Otherwise on 32-bit systems a 64-bit buffer view cannot be converted to
a pointer, which causes problems. This entails checking for too large
size values in a few places and casting to `Int` in a few more places
where we know a value has to fit into a native `Int`.