read_standard_header: improve error checks & messages
This is a fairly significant refactoring of the error checking aspect of
header parsing with a few high-level improvements. The first major
improvement is that this now extracts and checks the version string
first, then the checksum and only if those are valid does it try
extracting any other fields. This makes the error message a bit clearer:
previously you would get a cryptic complaint about an invalid octal
digit if you tried to parse something that wasn't a tar file, now at
least you get an "invalid version string for tar file" error.

This now also dumps the entire header data if there is any error parsing
the header, which helps with diagnosis. In addition, the error checking
and reporting for boundary conditions in fields which can be malformed
(i.e. octal integers) is much improved.

Last but not least, the test coverage of all of this is much increased.

(cherry picked from commit cbcb00b7b33c87792ab35b8325d14ee635d26581)