read_header_int: allow leading spaces (fix #111)
Some old versions of `tar` (e.g. Solaris) emitted integer fields with
leadings *spaces* which is technically contrary to the standard, but
seems harmless to support. This adds a fragment of this tarball:

    https://sparse.tamu.edu/MM/Oberwolfach/LF10.tar.gz

This is a collection of example sparse matrix files. The fragment is a
single header entry followed by one block of data and is a valid tarball
by itself with the idiosyncratic field formatting of Solaris `tar`.