handle extract of self hardlink (fix #127)
This is a weird construction: an entry for a file followed by a hardlink
to what that file as its target and with the same path. The hardlink
copies the content of its target and creates a file with the "new" path
and mode, which happens in this case to be the same path as the original
file. The effect of this construction is to change the permissions of an
previous file entry. We could diallow this since it's odd, but based on
the bug report it is something `tar` can generate and the other tarball
reading methods like `tree_hash` and `rewrite` already just work because
the logic is simply to copy the contents of an existing node with a new
path and mode and for a sane tree data structure, you can just overwrite
an arbitrary node. The file system is wonkier and the change in logic
here is merely to ensure that the old file isn't deleted too early.