test that extracting outside of dest dir is an error; also check harder
We have always prevented this for security reasons, but although we have
tested that the fancy attacks using symlinks are prevented, we haven't
been testing that the basic attack of extracting a relative or absolute
path outside of the tarball is prevented. This adds tests for that.

This also adds some additional paranoia in `Tar.extract` to make certain
that the final paths we end up extracting are not absolute and that they
don't use any relative path tricks to end up outside of the extraction
root directory.