Check for bad Windows paths prior to extraction; allow check on creation
Previously, we didn't bat an eye at extracting paths that are considered
bad or illegal on Windows. This explicitly checks for such paths before
attempting extraction on Windows. On all platforms it introduces the
`portable` flag for tarball creating operations, which does the same
checks and errors if you would have created a tarball that would error
upon extraction on Windows.