Fix some issues found by JET (#162)
The first problem is that if a skeleton file starts with zero blocks
then `read_standard_header` can return `nothing` which we didn't check
for; the fix is to check for it. We also add a test for that situation.

The second problem is that while writing header blocks, we wrap an array
view in an IOBuffer which we write to. However, the IOBuffer write
implementation can try to grow the vector that's passed in, which array
views do not support. This shouldn't happen since we ensure that the
buffer is at least 512 bytes and we should never write more than 512
blocks, but JET can't possibly know that. This refactors the code to
wrap the buffer directly instead of wrapping a view, which avoids this
issue since the buffer *can* be grown by the IOBuffer write code. No
test added since there isn't any way to actually trigger this error.

Finally, I add type upper bound on HashType <: SHA.SHA_CTX, which is
also something that JET was unhappy about.